[Refactor] change folder name
